analyzers are now classes; makefile now automatically generates boilerplate code for analyzers; razoraux files are now merged into RazorAnalyzer.cc
