+++ .gitignore
@@ -26,3 +26,4 @@ test.*
 *.tex
 *.gif
 *.png
+bin
