+++ Makefile
@@ -3,28 +3,29 @@ include Makefile.inc
 DIRS = python
 SRCDIR = src
 INCLUDEDIR = include
+ANADIR = analyzers
+BINDIR = bin
 INCLUDELIST= SimpleTable.h Linkdef.h
 
-SCRATCHDIR = /tmp/
+ANALYZERS = $(wildcard $(ANADIR)/*.cc)
+ANALYZERSH = $(addprefix $(INCLUDEDIR)/,$(notdir $(ANALYZERS:cc=h)))
+ANALYZERSOBJ = $(ANALYZERS:cc=o)
+RUNNERS = $(addprefix $(BINDIR)/Run,$(notdir $(basename $(ANALYZERS))))
+RUNNERSCC = $(addsuffix .cc,$(addprefix $(ANADIR),$(notdir RUNNERS)))
+UTILS = $(SRCDIR)/JetCorrectorParameters.cc $(SRCDIR)/SimpleJetCorrectionUncertainty.cc  $(SRCDIR)/JetCorrectionUncertainty.cc $(SRCDIR)/SimpleJetCorrector.cc $(SRCDIR)/FactorizedJetCorrector.cc $(SRCDIR)/SimpleJetResolution.cc $(SRCDIR)/BTagCalibrationStandalone.cc $(SRCDIR)/EnergyScaleCorrection_class.cc $(SRCDIR)/Hemisphere.cc $(SRCDIR)/Davismt2.cc $(SRCDIR)/RazorHelper.cc
+UTILSOBJ = $(UTILS:cc=o)
+EXECUTABLES = NormalizeNtuple SkimNtuple $(RUNNERS)
+HELPERSCRIPT = python/MakeAnalyzerCode.py
 
-AUX = $(wildcard $(SRCDIR)/RazorAux*.cc $(SRCDIR)/HggRazorAuxPhoton.cc $(SRCDIR)/Hemisphere.cc $(SRCDIR)/Davismt2.cc)
-ANALYSES = $(wildcard analyses/*.cc)
-JETCORR = $(SRCDIR)/JetCorrectorParameters.cc $(SRCDIR)/SimpleJetCorrectionUncertainty.cc  $(SRCDIR)/JetCorrectionUncertainty.cc $(SRCDIR)/SimpleJetCorrector.cc $(SRCDIR)/FactorizedJetCorrector.cc $(SRCDIR)/SimpleJetResolution.cc $(SRCDIR)/BTagCalibrationStandalone.cc $(SRCDIR)/EnergyScaleCorrection_class.cc
-JETCORROBJ = $(JETCORR:cc=o)
-EXECUTABLES = RazorRun NormalizeNtuple SkimNtuple
+.PHONY: clean all lxplus 
 
-.PHONY: clean all lxplus t3higgs
-
-t3higgs: $(addprefix $(SCRATCHDIR)/, $(EXECUTABLES))
-	mv $(addprefix $(SCRATCHDIR)/, $(EXECUTABLES)) .
-	@for d in $(DIRS); do (cd $$d; $(MAKE) $(MFLAGS) ); done
 all: $(EXECUTABLES)
-	@for d in $(DIRS); do (cd $$d; $(MAKE) $(MFLAGS) ); done
+#	@for d in $(DIRS); do (cd $$d; $(MAKE) $(MFLAGS) ); done
 lxplus: all
 
 clean:
 	@-rm $(EXECUTABLES)
-	@rm -f $(SRCDIR)/*.o
+	@rm -f $(SRCDIR)/*.o $(ANADIR)/*.o
 	@for d in $(DIRS); do (cd $$d; $(MAKE) $(MFLAGS) clean ); done
 
 $(INCLUDEDIR)/rootdict.cc:
@@ -33,29 +34,26 @@ $(INCLUDEDIR)/rootdict.cc:
 $(SRCDIR)/SimpleTable.o: $(SRCDIR)/SimpleTable.cc 
 	$(CXX) -c $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
 
-$(SRCDIR)/RazorHelper.o: $(SRCDIR)/RazorHelper.cc 
-	$(CXX) -c $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
-
 $(SRCDIR)/RazorEvents.o: $(SRCDIR)/RazorEvents.C $(INCLUDEDIR)/RazorEvents.h
 	$(CXX) $(SRCDIR)/RazorEvents.C $(CXXFLAGS) -I$(INCLUDEDIR) -c $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
 
 $(SRCDIR)/RazorAnalyzer.o: $(SRCDIR)/RazorEvents.o $(SRCDIR)/RazorAnalyzer.cc
 	$(CXX) $(SRCDIR)/RazorAnalyzer.cc $(CXXFLAGS) -I$(INCLUDEDIR) -c $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
 
-$(JETCORROBJ): %.o: %.cc
+$(UTILSOBJ): %.o: %.cc
 	$(CXX) -c $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS) $<
 
-$(SCRATCHDIR)/RazorRun: $(SRCDIR)/RazorEvents.o $(SRCDIR)/RazorAnalyzer.o $(SRCDIR)/RazorHelper.o $(ANALYSES) $(JETCORROBJ) $(AUX) $(SRCDIR)/RazorRun.cc
-	$(CXX) $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
+$(ANALYZERSOBJ): $(ANADIR)/%.o: $(ANADIR)/%.cc $(INCLUDEDIR)/%.h
+	$(CXX) -c $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS) $<
 
-$(SCRATCHDIR)/NormalizeNtuple: $(SRCDIR)/SimpleTable.o $(SRCDIR)/NormalizeNtuple.cc $(INCLUDEDIR)/rootdict.o
-	$(CXX) $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
+$(ANALYZERSH): 
+	$(HELPERSCRIPT) $(notdir $(basename $@))
 
-$(SCRATCHDIR)/SkimNtuple: $(SRCDIR)/SimpleTable.o $(SRCDIR)/SkimNtuple.cc $(INCLUDEDIR)/rootdict.o
-	$(CXX) $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
+$(RUNNERSCC): 
+	$(HELPERSCRIPT) $(notdir $(basename $($@:Run=)))
 
-RazorRun: $(SRCDIR)/RazorEvents.o $(SRCDIR)/RazorAnalyzer.o $(SRCDIR)/RazorHelper.o $(ANALYSES) $(JETCORROBJ) $(AUX) $(SRCDIR)/RazorRun.cc
-	$(CXX) $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
+$(RUNNERS): $(BINDIR)/Run%: $(SRCDIR)/RazorEvents.o $(SRCDIR)/RazorAnalyzer.o $(UTILSOBJ) $(ANADIR)/%.o $(SRCDIR)/Run%.cc
+	$(CXX) $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS) 
 
 NormalizeNtuple: $(SRCDIR)/SimpleTable.o $(SRCDIR)/NormalizeNtuple.cc $(INCLUDEDIR)/rootdict.o
 	$(CXX) $^ $(CXXFLAGS) -I$(INCLUDEDIR) $(LDFLAGS) $(LIBS) -o $@ $(CXX11FLAGS)
