+++ include/AnalyzerTemplate.txt
@@ -1,0 +1,12 @@
+#ifndef DEF_%ANALYZER%
+#define DEF_%ANALYZER%
+
+#include "RazorAnalyzer.h"
+
+class %ANALYZER%: public RazorAnalyzer {
+    public: 
+        %ANALYZER%(TTree *tree=0): RazorAnalyzer(tree) { }
+        void Analyze();
+};
+
+#endif
